+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/access.js
@@ -1,0 +1,9 @@
+module.exports = {
+	guests: false,
+	logged: true,
+	http:   true,
+	https:  true,
+	intro:  false,
+	index:  false,
+	groups: []
+}