+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/register.json/post.js
@@ -1,7 +1,9 @@
 ﻿module.exports = function(client, callback) {
+
 	security.register(client, function(err, user) {
 		if (user) client.context.data = { Result: "Ok" };
 		else client.context.data = { Result: "Error" };
 		callback();
 	});
+
 }