+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/regvalidation.json/post.js
@@ -1,5 +1,7 @@
 ﻿module.exports = function(client, callback) {
+
 	security.getUser(client, client.fields.Email, function(err, user) {
 		callback({ Email: !user });
 	});
+
 }