+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/signIn.json/access.js
@@ -1,0 +1,7 @@
+module.exports = {
+	guests: true,
+	logged: false,
+	http:   true,
+	https:  true,
+	groups: []
+}