+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/signIn.json/post.js
@@ -1,8 +1,10 @@
 ﻿module.exports = function(client, callback) {
+
 	security.signIn(client, function(isSuccess) {
 		if (isSuccess) client.context.data = { Result: "Ok" };
 		else client.context.data = { Result: "Error" };
 		if (client.fields.loginForm) client.redirect("/");
 		callback();
 	});
+
 }