+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/signOut.json/post.js
@@ -1,4 +1,6 @@
 ﻿module.exports = function(client, callback) {
+
+	client.context.data = { Result: "Ok" };
 	security.signOut(client, callback);
-	callback({ Result: "Ok" })
+
 }