+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/userInfo.json/get.js
@@ -1,4 +1,6 @@
 module.exports = function(client, callback) {
-    client.context.data = client.req.user ? client.req.user  : 'mde';
+
+    client.context.data = client.req.user ? client.req.user : {};
     callback();
+
 };